+++ armonaut/auth/models.py
@@ -12,8 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from sqlalchemy import Column, String, DateTime, Boolean
+from sqlalchemy import Column, String, DateTime, Boolean, Text, ForeignKey
 from sqlalchemy import sql
+from sqlalchemy.orm import relationship
 from armonaut.db import Model
 
 
@@ -23,9 +24,25 @@ class User(Model):
     email = Column(String(96), nullable=False, index=True, unique=True)
     password = Column(String(128), nullable=False)
 
-    is_staff = Column(Boolean, nullable=False)
     is_superuser = Column(Boolean, nullable=False)
+    is_staff = Column(Boolean, nullable=False)
 
     joined_date = Column(DateTime, nullable=False, server_default=sql.func.now())
     password_date = Column(DateTime, nullable=False, server_default=sql.func.now())
     last_login = Column(DateTime, nullable=False, server_default=sql.func.now())
+
+    accounts = relationship('Account', back_populates='user')
+
+
+class Account(Model):
+    __tablename__ = 'accounts'
+
+    username = Column(String(255), nullable=False, index=True)
+    access_token = Column(Text, nullable=False)
+    refresh_token = Column(Text, nullable=False)
+
+    user = relationship('User', back_populates='accounts', uselist=False)
+    user_id = Column(
+        ForeignKey('users.id', onupdate='CASCADE', ondelete='CASCADE'),
+        nullable=False
+    )