+++ armonaut/webhooks/views.py
@@ -11,10 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-from zope.interface import Interface
-
-
-class IWebhookVerifierService(Interface):
-    def verify_request(self, request, secret):
-        raise NotImplementedError()