+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
 
-- Added GitHub webhook verifier service.
+- Added Projects, Roles, and Accounts.
+- Added GitLab webhook service.
+- Added BitBucket webhook service.
+- Added GitHub webhook service.
 - Added basic login template.
 - Added BasicAuth and Session-based authentication.
 - Added password reset token service.
@@ -29,7 +34,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Gulp configuration for compiling static resources.
 - Added Dockerfile and Docker-Compose setup for testing and development.
 
-### Tools
+### Developer Tools
 
 - Added `armonaut db ...` tools for managing migrations via Alembic.
 - Added `armonaut ctrl ...` tool for controlling `docker-compose` commands and